requirements.txt
python-dotenv
scikit-learn
numpy
xmltodict
google-genai
firebase-admin